+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
+- Fixed visual issue where resizing elements caused the text inside of the element to be selected.
+- Fixed issue where dragging elements did not disable the content editing mode.
+- Fixed issue where elements with `box-sizing: content-box` and padding were not resized correctly.
 - Added status badges to the README.md file.
 - Added a contributor section to the README.md file.