+++ CHANGELOG.md
@@ -11,11 +11,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Added a console warning when users try to resize a rotated element: "Remarklet does not yet support resizing rotated elements."
 - Added status badges to the README.md file.
 - Added a contributor section to the README.md file.
+- Added a stylesheet utility module to handle creating a stylesheet for the library that can receive updates based on state changes.
 
 ### Changed
 
 - Resizing will now only set the width or height of the element during the action, not both at the same time.
-- Fixed visual issue where resizing elements caused the text inside of the element to be selected.
+- Fixed visual issue where resizing elements caused the text inside of elements to be selected.
 - Fixed issue where dragging elements did not disable the content editing mode.
 - Fixed issue where elements with `box-sizing: content-box` and padding were not resized correctly.